refactor(DynamoStore): move the remove_empty_strings method to the new DictUtils.py (#33)
